+++ src/javascripts/components/boardCard/boardCard.scss
@@ -2,3 +2,7 @@
   height: 300px;
   width: 300px;
 }
+
+.singleBoard {
+  *{pointer-events: none;}
+}
