+++ src/javascripts/components/buildTheSite/buildTheSite.scss
@@ -1,0 +1,4 @@
+.header {
+  margin: 20px;
+  padding: 20px;
+}
