+++ src/javascripts/components/pinCard/pinCard.js
@@ -1,24 +1,17 @@
-// import './pinCard.scss';
-// // import pinsData from '../../helpers/data/pinsData';
-// // import b from '../boards/board';
+import './pinCard.scss';
 
-// const pinCard = (pins) => {
-//   pins.forEach((pin) => {
-//     console.log(pin);
-//     let domString = '';
-//     domString += `
-//     <div class="card col-4">
-//     <h5 class="card-title text-center">${pins.name}</h5>
-//     <div class="card-body">
-//     <img src=${pins.imageUrl} class="card-img-top" alt="...">
-//     <p class="card-text text-center">${pins.description}</p>
-//     </div>
-//     </div>`;
-//     domString = `
-//     <div class="card col-4">
-//     <h5 class="card-title text-center">EMPTY</h5>`;
-//     return domString;
-//   });
-// };
+const printPinCards = (pin) => {
+  let domString = '';
+  domString += `
+  <div class="card col-3 single-pin" id="${pin.id}">
+  <h5 class="card-title">${pin.name}</h5>
+    <img src=${pin.imageUrl} class="card-img-top pinImg" alt="Image of ${pin.name}">
+    <div class="card-body">
+      <p>${pin.description}</p>
+    </div>
+  </div>
+  `;
+  return domString;
+};
 
-// export default { pinCard };
+export default { printPinCards };
